Fix 16-bit integer overflow for entry IDs in C64 client
Entry IDs exceeding 65535 (16-bit max) were wrapping around due to
using 'int' type on the C64, causing wrong entries to be loaded when
browsing categories with high ID ranges (Discmags, Graphics, etc.).

For example, Discmags entry ID 218475 wrapped to 21867 (a Games entry),
causing the client to load Games/CSDB content instead of Discmags/CSDB.

Changes:
- Changed item_ids array from int to long (32-bit)
- Changed atoi() to atol() for parsing entry IDs
- Updated run_entry() and fetch_info() to use long parameter
- Updated sprintf format strings to use %ld

Also added debug logging in server to trace LIST grouped entries,
and fixed cross-category grouping in handleGroupedResults().

+++ c64client/src/main.c
@@ -42,7 +42,7 @@ static bool connected = false;
 // Menu/list state
 #define MAX_ITEMS 20
 static char item_names[MAX_ITEMS][32];
-static int  item_ids[MAX_ITEMS];
+static long item_ids[MAX_ITEMS];
 static int  item_count = 0;
 static long total_count = 0;
 static int  cursor = 0;
@@ -368,7 +368,7 @@ void load_entries(const char *category, int start)
         if (name)
         {
             *name++ = 0;
-            item_ids[item_count] = atoi(id_str);
+            item_ids[item_count] = atol(id_str);
 
             // Find category (next |)
             char *cat = strchr(name, '|');
@@ -425,12 +425,12 @@ void load_entries(const char *category, int start)
 }
 
 // Run selected entry
-void run_entry(int id)
+void run_entry(long id)
 {
     print_status("running...");
 
     char cmd[32];
-    sprintf(cmd, "RUN %d", id);
+    sprintf(cmd, "RUN %ld", id);
     send_command(cmd);
 
     read_line();  // "OK Running xxx" or "ERR xxx"
@@ -478,7 +478,7 @@ void do_search(const char *query, int start)
         if (name)
         {
             *name++ = 0;
-            item_ids[item_count] = atoi(id_str);
+            item_ids[item_count] = atol(id_str);
 
             // Find end of name (next |)
             char *end = strchr(name, '|');
@@ -573,7 +573,7 @@ void do_adv_search(int start)
         if (name)
         {
             *name++ = 0;
-            item_ids[item_count] = atoi(id_str);
+            item_ids[item_count] = atol(id_str);
 
             // Find category (next |)
             char *cat = strchr(name, '|');
@@ -670,7 +670,7 @@ void do_releases(const char *category, const char *title, int start)
         if (grp)
         {
             *grp++ = 0;
-            item_ids[item_count] = atoi(id_str);
+            item_ids[item_count] = atol(id_str);
 
             // Find year (next |)
             char *year = strchr(grp, '|');
@@ -720,12 +720,12 @@ void do_releases(const char *category, const char *title, int start)
 }
 
 // Fetch info for an entry
-bool fetch_info(int id)
+bool fetch_info(long id)
 {
     print_status("loading info...");
 
     char cmd[32];
-    sprintf(cmd, "INFO %d", id);
+    sprintf(cmd, "INFO %ld", id);
     send_command(cmd);
     read_line();  // "OK" or "ERR ..."
